reductionUMAP

Run UMAP to project samples into 2D space using pairwise distances

reductionUMAP(object, seed = 42, method = "naive", n.neighbors = 15,
  n.components = 2, metric = "manhattan", verbose = TRUE,
  n.epochs = 200, min.dist = 0.1, spread = 1, set.op.mix.ratio = 1,
  local.connectivity = 1L, negative.sample.rate = 5L)

Arguments

object
Cookie object

seed
see number. default is 42

method
could be "naive" or "umap-learn". If choose "umap-learn", user may need to install python package umap-learn (https://pypi.org/project/umap-learn/)

n.neighbors
integer; number of nearest neighbors

n.components
integer; dimension of target (output) space

metric
character or function; determines how distances between data points are computed. When using a string, available metrics are: euclidean, manhattan. Other available generalized metrics are: cosine, pearson, pearson2. Note the triangle inequality may not be satisfied by some generalized metrics, hence knn search may not be optimal. When using metric.function as a function, the signature must be function(matrix, origin, target) and should compute a distance between the origin column and the target columns

verbose
logical or integer; determines whether to show progress messages

n.epochs
integer; number of iterations performed during layout optimization

min.dist
numeric; determines how close points appear in the final layout

spread
numeric; used during automatic estimation of a/b parameters.

set.op.mix.ratio
numeric in range [0,1]; determines who the knn-graph is used to create a fuzzy simplicial graph

local.connectivity
numeric; used during construction of fuzzy simplicial set

negative.sample.rate
integer; determines how many non-neighbor points are used per point and per iteration during layout optimization